Pulls filename settings into it's own component
