+++ src/settings/settings.tsx
@@ -1,12 +1,12 @@
 import { type App, Notice, PluginSettingTab, Setting, moment } from 'obsidian';
 import { createRoot, type Root } from 'react-dom/client';
-import { useEffect, useMemo, useRef, useState } from 'react';
-import { debounce } from 'mini-debounce';
+import { useDebounce } from 'src/util/useDebounce';
 
 import type ScribePlugin from 'src';
 
-import { formatFilenamePrefix } from 'src/util/filenameUtils';
 import { LLM_MODELS } from 'src/util/openAiUtils';
+
+import { FileNameSettings } from './components/FileNameSettings';
 
 export enum TRANSCRIPT_PLATFORM {
   assemblyAi = 'assemblyAi',
@@ -208,131 +208,17 @@ export class ScribeSettingsTab extends PluginSettingTab {
 
   saveSettings() {
     this.plugin.saveSettings();
-    new Notice('Scribe: ✅ Settings saved');
   }
 }
 
 const ScribeSettings: React.FC<{ plugin: ScribePlugin }> = ({ plugin }) => {
-  const [noteFilenamePrefix, setNoteFilenamePrefix] = useState(
-    plugin.settings.noteFilenamePrefix,
-  );
-  const [recordingFilenamePrefix, setRecordingFilenamePrefix] = useState(
-    plugin.settings.recordingFilenamePrefix,
-  );
-  const [dateFilenameFormat, setDateFilenameFormat] = useState(
-    plugin.settings.dateFilenameFormat,
-  );
-  const isDateInPrefix =
-    (noteFilenamePrefix || '').includes('{{date}}') ||
-    (recordingFilenamePrefix || '').includes('{{date}}');
-
   const savedNotif = useDebounce(() => {
     plugin.saveSettings();
-  }, 500);
+  }, 700);
 
   return (
     <div>
-      <h2>File name properties</h2>
-      <SettingsItem
-        name="Transcript filename prefix"
-        description="This will be the prefix of the note filename, use {{date}} to include the date"
-        control={
-          <input
-            type="text"
-            placeholder="scribe-"
-            value={noteFilenamePrefix}
-            onChange={(e) => {
-              setNoteFilenamePrefix(e.target.value);
-              plugin.settings.noteFilenamePrefix = e.target.value;
-              savedNotif();
-            }}
-          />
-        }
-      />
-      <SettingsItem
-        name="Audio recording filename prefix"
-        description="This will be the prefix of the audio recording filename, use {{date}} to include the date"
-        control={
-          <input
-            type="text"
-            placeholder="scribe-"
-            value={recordingFilenamePrefix}
-            onChange={(e) => {
-              setRecordingFilenamePrefix(e.target.value);
-              plugin.settings.recordingFilenamePrefix = e.target.value;
-              savedNotif();
-            }}
-          />
-        }
-      />
-      <SettingsItem
-        name="Date format"
-        description="This will only be used if {{date}} is in the transcript or audio recording filename prefix above."
-        control={
-          <div>
-            <input
-              type="text"
-              placeholder="YYYY-MM-DD"
-              disabled={!isDateInPrefix}
-              value={dateFilenameFormat}
-              onChange={(e) => {
-                setDateFilenameFormat(e.target.value);
-                plugin.settings.dateFilenameFormat = e.target.value;
-                savedNotif();
-              }}
-            />
-          </div>
-        }
-      />
-      {isDateInPrefix && (
-        <div>
-          <p>
-            {formatFilenamePrefix(`${noteFilenamePrefix}`, dateFilenameFormat)}
-            filename
-          </p>
-          <p>
-            {formatFilenamePrefix(
-              `${recordingFilenamePrefix}`,
-              dateFilenameFormat,
-            )}
-            filename
-          </p>
-        </div>
-      )}
+      <FileNameSettings plugin={plugin} saveSettings={savedNotif} />
     </div>
   );
-};
-
-const SettingsItem: React.FC<{
-  name: string;
-  description: string;
-  control: React.ReactNode;
-}> = ({ name, description, control }) => {
-  return (
-    <div className="setting-item">
-      <div className="setting-item-info">
-        <div className="setting-item-name">{name}</div>
-        <div className="setting-item-description">{description}</div>
-      </div>
-      <div className="setting-item-control">{control}</div>
-    </div>
-  );
-};
-
-const useDebounce = (callback: () => void, timeMs: number): (() => void) => {
-  const ref = useRef<() => void>();
-
-  useEffect(() => {
-    ref.current = callback;
-  }, [callback]);
-
-  const debouncedCallback = useMemo(() => {
-    const func = () => {
-      ref.current?.();
-    };
-
-    return debounce(func, timeMs);
-  }, [timeMs]);
-
-  return debouncedCallback;
 };
